--- PennMobile/src/main/java/com/pennapps/labs/pennmobile/LaundryFragment2.kt
@@ -1,141 +1,0 @@
-package com.pennapps.labs.pennmobile
-
-import StudentLife_rf2
-import android.content.Context
-import android.content.SharedPreferences
-import android.os.Bundle
-import androidx.preference.PreferenceManager
-import android.view.*
-import androidx.coordinatorlayout.widget.CoordinatorLayout
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentTransaction
-import androidx.fragment.app.activityViewModels
-import androidx.recyclerview.widget.LinearLayoutManager
-import com.pennapps.labs.pennmobile.adapters.LaundryRoomAdapter
-import com.pennapps.labs.pennmobile.classes.LaundryRoom
-import com.pennapps.labs.pennmobile.classes.LaundryUsage
-import com.pennapps.labs.pennmobile.components.collapsingtoolbar.ToolbarBehavior
-import com.pennapps.labs.pennmobile.databinding.FragmentLaundryBinding
-import com.pennapps.labs.pennmobile.utils.Utils
-import com.pennapps.labs.pennmobile.viewmodels.LaundryViewModel
-import kotlinx.android.synthetic.main.loading_panel.*
-import java.util.*
-
-class LaundryFragment2 : Fragment() {
-    private lateinit var mActivity: MainActivity
-
-    private lateinit var mStudentLife: StudentLife_rf2
-    private lateinit var mContext: Context
-
-    private lateinit var sharedPreferences: SharedPreferences
-
-    // list of favorite laundry rooms
-    private var laundryRooms = ArrayList<LaundryRoom>()
-    // data for laundry room usage
-    private var roomsData: ArrayList<LaundryUsage> = ArrayList()
-
-    private var mAdapter: LaundryRoomAdapter? = null
-
-    private var _binding : FragmentLaundryBinding? = null
-    private val binding get() = _binding!!
-
-    private val laundryViewModel : LaundryViewModel by activityViewModels()
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-
-        mStudentLife = MainActivity.studentLifeInstanceRf2
-        mActivity = activity as MainActivity
-        mContext = mActivity
-        sharedPreferences = PreferenceManager.getDefaultSharedPreferences(mActivity)
-    }
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View {
-        _binding = FragmentLaundryBinding.inflate(inflater, container, false)
-        val view = binding.root
-
-        initAppBar()
-
-        binding.favoriteLaundryList.layoutManager = LinearLayoutManager(mContext)
-        binding.laundryMachineRefresh.setOnRefreshListener {
-            updateMachines()
-        }
-        binding.laundryMachineRefresh.setColorSchemeResources(R.color.color_accent, R.color.color_primary)
-
-        updateMachines()
-        return view
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        mActivity.removeTabs()
-        mActivity.setTitle(R.string.laundry)
-
-        mAdapter = LaundryRoomAdapter(mContext, laundryRooms, roomsData, false)
-        binding.favoriteLaundryList.adapter = mAdapter
-
-        loadingPanel?.visibility = View.VISIBLE
-
-        laundryViewModel.favoriteRooms.observe(viewLifecycleOwner) { favorites ->
-            binding.laundryMachineRefresh.isRefreshing = false
-
-            laundryRooms.clear()
-            roomsData.clear()
-
-            laundryRooms.addAll(favorites.favoriteRooms)
-            roomsData.addAll(favorites.roomsData)
-
-            for (pos in 0 until LaundryViewModel.maxNumRooms) {
-                mAdapter!!.notifyItemChanged(pos)
-            }
-
-            loadingPanel?.visibility = View.GONE
-            binding.laundryHelpText.visibility = View.INVISIBLE
-        }
-
-    }
-
-    private fun getOnline() : Boolean {
-        //displays banner if not connected
-        if (!isOnline(context)) {
-            binding.internetConnectionLaundry.setBackgroundColor(resources.getColor(R.color.darkRedBackground))
-            binding.internetConnectionMessageLaundry.text = getString(R.string.internet_error)
-            binding.internetConnectionLaundry.visibility = View.VISIBLE
-            binding.laundryHelpText.visibility = View.INVISIBLE
-            binding.laundryMachineRefresh.isRefreshing = false
-            loadingPanel?.visibility = View.GONE
-            return false
-        }
-
-        binding.internetConnectionLaundry.visibility = View.GONE
-        return true
-    }
-    private fun updateMachines() {
-        if (!getOnline()) {
-            return
-        }
-        mActivity.mNetworkManager.getAccessToken {
-            val bearerToken = "Bearer " + sharedPreferences
-                .getString(getString(R.string.access_token), "").toString()
-            laundryViewModel.getFavorites(mStudentLife, bearerToken)
-        }
-    }
-
-    private fun initAppBar() {
-        (binding.appbarHome.layoutParams as CoordinatorLayout.LayoutParams).behavior = ToolbarBehavior()
-        binding.titleView.text = getString(R.string.laundry)
-        binding.dateView.text = Utils.getCurrentSystemTime()
-        binding.laundryPreferences.setOnClickListener {
-            val fragmentManager = mActivity.supportFragmentManager
-            fragmentManager.beginTransaction()
-                .replace(R.id.content_frame, LaundrySettingsFragment2())
-                .addToBackStack("Laundry Settings Fragment")
-                .setTransition(FragmentTransaction.TRANSIT_FRAGMENT_OPEN)
-                .commit()
-        }
-    }
-}
